Changed Database SQL value field to floatl

+++ Database.sql
@@ -2,7 +2,7 @@ CREATE TABLE `sensors` (
   `id` int(11) unsigned NOT NULL AUTO_INCREMENT,
   `user_id` int(11) NOT NULL,
   `sensor_id` int(11) NOT NULL,
-  `value` int(11) NOT NULL,
+  `value` float NOT NULL,
   `timestamp` datetime NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
